Change option --scriptname to --shortname and change help text.
Option --scriptname did nothing, because the program expected --shortname instead. --scriptname is far more confusing, given that it's intended as the value that lands in the __SHORTFILE__ macro, so it has been renamed.

Improve the help text regarding that option; also fix the wording for what -O help does and add an example of how to use -O.