Remove llGetEnv tests from tests; don't generate list in list
llGetEnv() is not computable now, so remove it from the computable functions unit test.

The test generated lists inside a list, and after the llGetEnv() change, that caused a weird side effect that hasn't been investigated (see test result of previous commit). To be on the safe side, take all list-generating results out of the list, into their own llSetPrimitiveParams call.