Fix wrong output leading to incorrect tokenization of minus signs
When the tree has a unary minus node (NEG) whose child is a product node (*), and the left operand of the product node starts with a minus sign but is not a NEG node, this produced at least two minus signs in sequence without any spaces.

Normally, OptSigns hides this problem, but when it is disabled, or when the left factor is a pre-decrement (--V) node, the problem is visible.

Fix by creating a function that detects all kinds of leading minus signs, and use it in place of the comparison with NEG.

Fixes #31. Reported by @KrsityKu, who also provided a repro.